Fix mac build

+++ Core/HLE/sceMpeg.cpp
@@ -838,7 +838,7 @@ u32 sceMpegRingbufferPut(u32 ringbufferAddr, u32 numPackets, u32 available)
 	if (ringbuffer.callback_addr) {
 		PostPutAction *action = (PostPutAction *) __KernelCreateAction(actionPostPut);
 		action->setRingAddr(ringbufferAddr);
-		u32 args[3] = {ringbuffer.data, numPackets, ringbuffer.callback_args};
+		u32 args[3] = {(u32)ringbuffer.data, numPackets, (u32)ringbuffer.callback_args};
 		__KernelDirectMipsCall(ringbuffer.callback_addr, action, false, args, 3, false);
 	} else {
 		ERROR_LOG(HLE, "sceMpegRingbufferPut: callback_addr zero");
